Python Code/GP_API/requirements.txt
Cython==0.29.28
Flask==2.1.1
azure-cognitiveservices-formrecognizer==0.1.1
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.28
backcall==0.2.0
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-core==1.7.2
google-cloud-vision==2.7.2
google-cloud-translate==2.0.1
googleapis-common-protos==1.56.0
gunicorn==20.1.0
msrest==0.6.21
numpy==1.21.2
opencv-python==4.5.5.64
packaging==21.3
pandas==1.3.2
parso==0.8.3
pickleshare==0.7.5
pip==22.0.4
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
wheel==0.36.2
cmake==3.22.4
face_recognition==1.3.0
face_recognition-models==0.3.0
dlib==19.24.0
